Изменения Int а long,пару доделок todo

+++ src/main/java/vsu/labs/crypto/entity/JpaRepository/TestRepository.java
@@ -3,5 +3,5 @@ package vsu.labs.crypto.entity.JpaRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vsu.labs.crypto.entity.Test;
 
-public interface TestRepository extends JpaRepository<Test,Integer> {
+public interface TestRepository extends JpaRepository<Test,Long> {
 }
